% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ah02_extract_user_info.R
\name{extract_export_time}
\alias{extract_export_time}
\title{Extracts the Date at which the Fitness Data was Exported}
\usage{
extract_export_time(fitness_data)
}
\arguments{
\item{fitness_data}{Fitness data created with
[read_apple_health](read_apple_health)}
}
\description{
Extracts the Date at which the Fitness Data was Exported
}
\examples{
library(dplyr)
data_file <- system.file("example_data/Export.zip", package = "healthieR")
read_apple_health(data_file) \%>\% extract_export_time
}
\seealso{
Other extraction_functions: \code{\link{extract_personal_info}},
  \code{\link{extract_records}},
  \code{\link{extract_steps}}, \code{\link{extract_weight}}
}
