% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ah02_extract_user_info.R
\name{extract_personal_info}
\alias{extract_personal_info}
\title{Extracts information about the user}
\usage{
extract_personal_info(fitness_data)
}
\arguments{
\item{fitness_data}{Fitness data created with
[read_apple_health](read_apple_health)}
}
\description{
This function extracts certain chacterisitcs (see \emph{Extracted Information})
about the athlete if they are present in the data.
}
\section{Extracted Information}{

\itemize{
\item Date of Birth (Column \code{date_of_birth})
\item Sex (Column \code{sex})
\item Blood Type (Column \code{blood_type})
\item Fitzpatrick Skin Type (Column \code{skin_type})
}
}

\examples{
library(dplyr)
data_file <- system.file("example_data/Export.zip", package = "healthieR")
read_apple_health(data_file) \%>\% extract_personal_info
}
\seealso{
Other extraction_functions: \code{\link{extract_export_time}},
  \code{\link{extract_records}},
  \code{\link{extract_steps}}, \code{\link{extract_weight}}
}
