% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ah03_extract_records.R
\name{extract_records}
\alias{extract_records}
\title{Extracts All Records from the Fitness Data}
\usage{
extract_records(fitness_data)
}
\arguments{
\item{fitness_data}{Fitness data created with
[read_apple_health](read_apple_health)}
}
\description{
Extracts All Records from the Fitness Data
}
\examples{
library(dplyr)
data_file <- system.file("example_data/Export.zip", package = "healthieR")
read_apple_health(data_file) \%>\% extract_records
}
\seealso{
Other extraction_functions: \code{\link{extract_export_time}},
  \code{\link{extract_personal_info}},
  \code{\link{extract_steps}}, \code{\link{extract_weight}}
}
