% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ah01_read_data.R
\name{read_apple_health}
\alias{read_apple_health}
\title{Read Fitness Data From Apple Health}
\usage{
read_apple_health(export_zip)
}
\arguments{
\item{export_zip}{Path to the "Export.zip" file from the Apple Health App.}
}
\description{
This function reads XML data dumps from Apple Health.
}
\examples{
zip_file <- system.file("example_data/Export.zip", package = "healthieR")
health_data <- read_apple_health(zip_file)
print(health_data)
}
